Inline expansions and cosmetic knots
Introduces the util.inlineExpansion helper for adding links that, when
clicked, expand into additional text (and can have side effects).  Used
once during Cassia's dream to eliminate a duplication of text.

This change breaks our test explorer, so I've introduced a concept
called "cosmetic knots" which is any knot representing a change _within_
a passage instead of a motion between passages. The test runner
automatically determines which knots are cosmetic by observing whether
the passage name has changed.  Snowman can't restore cosmetic knots
directly, so to restore them we rewind to the last non-cosmetic knot (a
passage navigation) and then click the sequence of links that lead back
to the cosmetic knot.  I've tested this with an exploration of the dream
sequence, which contains exactly two paths right now - one in which the
expando is clicked, and one in which it's skipped.

I've torn out some other tests which are likely to get in the way while
we're writing a lot of content.